list 검색기능 개발중 querydsl 사용
할것 :
회원가입시 메시지 출력 깔끔하게 만들기(보류중)
소재지 파악해서 특정 지역만 보여주는거
데이터 넣을때 좀 깔끔하게 넣게 필터링 빡씨게
html 부분 깔끔하게

+++ src/main/java/com/myspring/springmaster/dataAccess/entity/House.java
@@ -19,7 +19,7 @@ public class House {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "name", length = 100)
     private String name;
@@ -54,7 +54,7 @@ public class House {
     private List<HouseDetail> houseDetails;
 
     @Builder
-    public House(Integer id, String name, String address, BigDecimal latitude, BigDecimal longitude, String status, String moveInDate, String applyStartDate, String applyEndDate, List<HouseDetail> houseDetails){
+    public House(Long id, String name, String address, BigDecimal latitude, BigDecimal longitude, String status, String moveInDate, String applyStartDate, String applyEndDate, List<HouseDetail> houseDetails){
         this.id = id;
         this.name = name;
         this.address = address;
